fix(places): implement shared data source to fix marker selection issues
